♻️ refactor: convert Auth.js v5 tests to Cucumber BDD format
